requirements.txt
flask==2.0.1
torch==1.9.0
torchvision==0.10.0
opencv-python==4.5.3.56
Pillow==8.3.2
numpy==1.21.2 